backend/pyproject.toml
[project]
name = "bh-cats-backend"
version = "0.1.0"
description = "BH Cats backend"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "pydantic-settings>=2.3,<3.0",
  "python-jose[cryptography]>=3.3,<4.0",
  "passlib[bcrypt]>=1.7,<2.0",
  "python-multipart>=0.0.9,<1.0",
  "pillow>=11.0,<12.0",
  "email-validator>=2.2,<3.0",
  "httpx>=0.28,<1.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-cov>=5.0,<6.0"
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
